+++ app/src/main/java/com/tgyuu/baekyoung_i/auth/navigation/AuthNavigation.kt
@@ -10,6 +10,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.navOptions
+import com.tgyuu.baekyoung_i.auth.AuthRoute
 import com.tgyuu.baekyoung_i.auth.AuthScreen
 
 const val authNavigationRoute = "auth_route"
@@ -22,26 +23,4 @@ fun NavGraphBuilder.authScreen(navigateToHome: () -> Unit) {
     composable(route = authNavigationRoute) {
         AuthRoute(navigateToHome)
     }
-}
-
-@Composable
-internal fun AuthRoute(navigateToHome: () -> Unit) {
-    var id by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-
-    val idChanged: (String) -> Unit = { textValue ->
-        id = textValue
-    }
-
-    val passwordChanged: (String) -> Unit = { textValue ->
-        password = textValue
-    }
-
-    AuthScreen(
-        id = id,
-        password = password,
-        idChanged = idChanged,
-        passwordChanged = passwordChanged,
-        navigateToHome = navigateToHome,
-    )
 }
