+++ app/src/main/java/com/tgyuu/baekyoung_i/consulting/ConsultingScreen.kt
@@ -17,6 +17,11 @@ import com.tgyuu.designsystem.component.BaekyoungTopAppBar
 import com.tgyuu.designsystem.theme.BaekyoungTheme
 
 @Composable
+internal fun ConsultingRoute() {
+    ConsultingScreen()
+}
+
+@Composable
 fun ConsultingScreen() {
     Box(
         modifier = Modifier
