+++ app/src/main/java/com/tgyuu/baekyoung_i/consulting/navigation/ConsultingNavigation.kt
@@ -6,6 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.navOptions
+import com.tgyuu.baekyoung_i.consulting.ConsultingRoute
 import com.tgyuu.baekyoung_i.consulting.ConsultingScreen
 
 const val consultingNavigationRoute = "consulting_route"
@@ -18,9 +19,4 @@ fun NavGraphBuilder.consultingScreen() {
     composable(route = consultingNavigationRoute) {
         ConsultingRoute()
     }
-}
-
-@Composable
-internal fun ConsultingRoute() {
-    ConsultingScreen()
 }
