+++ app/src/main/java/com/tgyuu/baekyoung_i/etc/EtcScreen.kt
@@ -15,6 +15,11 @@ import com.tgyuu.designsystem.component.BaekyoungTopAppBar
 import com.tgyuu.designsystem.theme.BaekyoungTheme
 
 @Composable
+internal fun EtcRoute() {
+    EtcScreen()
+}
+
+@Composable
 fun EtcScreen() {
     Column(
         modifier = Modifier
