+++ app/src/main/java/com/tgyuu/baekyoung_i/etc/navigation/EtcNavigation.kt
@@ -6,6 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.navOptions
+import com.tgyuu.baekyoung_i.etc.EtcRoute
 import com.tgyuu.baekyoung_i.etc.EtcScreen
 
 const val etcNavigationRoute = "etc_route"
@@ -18,9 +19,4 @@ fun NavGraphBuilder.etcScreen() {
     composable(route = etcNavigationRoute) {
         EtcRoute()
     }
-}
-
-@Composable
-internal fun EtcRoute() {
-    EtcScreen()
 }
