+++ app/src/main/java/com/tgyuu/baekyoung_i/home/navigation/HomeNavigation.kt
@@ -6,6 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.navOptions
+import com.tgyuu.baekyoung_i.home.HomeRoute
 import com.tgyuu.baekyoung_i.home.HomeScreen
 
 const val homeNavigationRoute = "home_route"
@@ -18,9 +19,4 @@ fun NavGraphBuilder.homeScreen() {
     composable(route = homeNavigationRoute) {
         HomeRoute()
     }
-}
-
-@Composable
-internal fun HomeRoute() {
-    HomeScreen()
 }
